+++ spec/spec_helper.rb
@@ -4,6 +4,7 @@ require "simplecov"
 SimpleCov.start do
   add_filter "/spec/"
   enable_coverage :branch
+  minimum_coverage line: 100, branch: 100
 end
 
 require "dedupe_requests"